Add subcommand CLI.
Remove urbansim_compile and urbansim_serve and replace them
with a top level urbansim command that takes subcommands.
Currently has compile and serve commands.

The interface is built using cliff: https://cliff.readthedocs.org
I added that to the install requirements along with bottle, which
is required for serve. As another project I'll move the server
dependencies out to an extra_requires argument so they aren't installed
for pure modeling.