Add _FakeRegressionResults for wrapping saved model coefficients.
A fit RegressionModel stores a statsmodels RegressionResults subclass
and uses that for prediction. (And users can look at the statistics of
the model fit.) For the purposes of storing models on disk all we can
save is the model expression and the fit coefficients.
_FakeRegressionResults can wrap those to provide a RegressionResults
stand in when loading a fit model from a file.