killing templates - goodbye templates!
For a couple of months I made the ill-advised decision to create models using model templates.  This were hard to write as models if a little more transparent as to what the code was actually doing.  Thanks mostly to the work of Matt Davis, we've got a well specified API that is well documented enough that it seems to be transparent enough to diagnose most strange behavior in the data.  At this time I'm making the bold move to remove templates from the repo, lest anyone begin to use them.  Thankfully, this also puts to rest what became the urbansim/urbansim/urbansim directory hierarchy (ouch).  Instead models are stored in urbansim/models.