Change return values of mnl_estimate
Now returns a dictionary of log likelihood values
(keys: 'null', 'convergence', 'ratio') and a DataFrame with
columns 'Coefficient', 'Std. Error', and 'T-Score'. The
DataFrame is not given any special index in mnl_estimate,
but when it is used with MNLLocationChoiceModel it gets an
index from the design matrix column names return by patsy,
which are derived from the model patsy expression.