segmented regression estimation is now working
This appears to be working with my workflow where I edit the yaml directly.  The biggest issue was when using numpy types as the segmented keys (when building_type_ids are ints).  Also edited the yamlmodelrunner to peek into the config file and choose the segmented or non-segmented model based on the model_type.  I will definitely be making these segmented models most of the time so I'm not sure if the forking strategy is really necessary.