mix of change to developer and lcms
Unfortunately I didn't check in my developer changes before I started working on debug info in the regressions and pdfs.  So this push includes both.

At any rate, this is a basically working developer model.

As well as version of yamlmodelrunner which returns debug information for the regressions and simulations.