Add .probabilities methods in LCM
On a basic LCM the .probabilities method returns an array of
probabilities and a Series of the alternatives those probabilities
are associated with. Right now those alternatives corresponding
to exactly the alternatives given because we calculate probabilities
for only one chooser, but that could change in the future.

On segmented LCMs and LCM groups there are two new methods.
.probabilities returns a dictionary where the keys are segment
names from the choosers table and the values are arrays of
probabilities for each segment. The .summed_probabilities methods
return one array of probabilities calculated by taking the
individual segment probabilities, normalizing them, and
multiplying by the number of members in the segment.