the two tests have been "fixed"
the second test was right this time and wrong before and was changed because the default settings changed.

the first test is difficult to fix, but it makes sense that Pandas doesn't allow this anymore.  I will take a look at it when I get a chance but should probably not spend too much time on it right now.