Support for two probability modes
Added support for modes 'single_chooser' and 'full_product'
when calculating probabilities. These actually affect the merging
of choosers and alternatives into an interaction dataset.
In 'single_chooser' only the first chooser in the choosers table
(after filtering) is used to construct the merged interaction table.
This is the same behavior as UrbanSim's previous LCM class.
In 'full_prodct' mode all choosers are merged with all alternatives.