Check mode compatibility in DCMs
At the moment certain probability and choice modes are incompatible
so this checks for those pairings and raises exceptions.
In particular the 'single_chooser' probability mode must be used
with the 'aggregate' choice mode and the 'full_product' probability
mode must be used with the 'individual' choice mode.

This also disallows using post-interaction filters in 'full_product'
mode because the filters can cause there to be different numbers of
alternatives for each chooser, which breaks the MNL machinery.