fixed a problem with deck parking in the developer
most of this commit is to make sure we account for the amount of a building that is parking and the amount that is not when crediting rents for the building.

the old structure of the pro forma was to maxmize the parking format (surface, deck, underground) in the precomputation part of the pro forma.  now this is put off until actual rents and costs are known.  this allows more flexibility in how the comutation is done.

in this case it actually removed a bug where we were giving too much rent for space that is actually only devoting to parking, which made itself known in the simulations as redeveloping sururban office developments that had surface parking and turning them into dense deck parking office developments because although the cost of the parking was taken into account, it was also given an offsetting rent because it was being subtracted from the utilized space.

this ended up being a very pervasive change which made me think that, more than ever, the developer model needs to be written in some way that is not vectorized.  presumably cython or just C or possibly directly in Excel and then stored as a lookup table.