tests should be passing
Because the logic of the developer model has changed a bit - see previous commit, the results changed slightly in a couple of cases in the tests.  I took a look and everything seems as expected, though I'd feel more comfortable with a revision of the developer model that is not vectorized as everything would be far more transparent (I mean, argmaxes are not exactly the most transparent operation).  This would hurt performance but is likely to be worth it.

This also has one fix as we used to sure a vector of average sqft prices and now we store the entire set of inputs/outputs for use in the "pick" operation.