fixed boundary case in supply and demand
We protect against the boundary case of having no choosers in the predict method, but this gets circumvented and is still a problem with the supply and demand code.

I was hoping to protect against this outside of urbansim, but we dont't know how many choosers there are until the predict_filters gets run, and that happens in urbansim.